Extract inline props code
